+++ models/task.model.js
@@ -1,0 +1,22 @@
+
+const { DataTypes, UUID } = require("sequelize");
+
+module.exports = (sequelize, Sequelize) => {
+    const Task = sequelize.define("Task", {
+        id: {
+            primaryKey: true,
+            type: DataTypes.UUID,
+            defaultValue: DataTypes.UUIDV4,
+        },
+        title: {
+            type: DataTypes.STRING,
+            allowNull: false,
+            defaultValue: "U didn't pick a title stoopid",
+        },
+        completed: {
+            type: DataTypes.BOOLEAN,
+        },
+    });
+
+    return Task;
+};
