+++ server.js
@@ -1,46 +1,37 @@
 const express = require("express");
-const { Sequelize, DataTypes, UUID } = require("sequelize");
+const tasksRouter = require('./routes/task.routs');
 require("dotenv").config();
 const app = express();
 const port = 3000;
-app.use(express.static("public"))
-app.set("view engine", "ejs")
-const sequelize = new Sequelize(
-	`postgres://${process.env.DB_USERNAME}:${process.env.DB_PASSWORD}@${process.env.DB_URL}:${process.env.DB_PORT}/${process.env.DB_NAME}`
-);
-
-const Task = sequelize.define("Task", {
-	id: {
-		primaryKey: true,
-		type: DataTypes.UUID,
-		defaultValue: DataTypes.UUIDV4,
-	},
-	title: {
-		type: DataTypes.STRING,
-		allowNull: false,
-		defaultValue: "U didn't pick a title stoopid",
-	},
-	completed: {
-		type: DataTypes.BOOLEAN,
-	},
-});
+const db = require("./models/index");
 
 const main = async () => {
 	try {
-		await sequelize.authenticate();
+		await db.sequelize.authenticate();
 		console.log("Connection has been established successfully.");
 	} catch (error) {
 		console.error("Unable to connect to the database:", error);
 	}
 };
+app.use(tasksRouter);
 
-const tasksRouter = require('./routes/task-routs');
+
+db.sequelize.sync()
+	.then(() => {
+		console.log("Synced db.");
+	})
+	.catch((err) => {
+		console.log("Failed to sync db: " + err.message);
+	});
+
+app.get("/", (req, res) => {
+	res.send("Welcome");
+});
 
 app.listen(port, () => {
 	console.log(`Example app listening on port ${port}`);
 });
 
 
-app.use(tasksRouter);
 
 main();
