--- views/task.ejs
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <form method="POST" action="/tasks/new">
-      <label for="name"
-        >Name:
-        <input
-          type="text"
-          id="taskID"
-          name="taskName"
-          required
-          value=" task new"
-        />
-      </label>
-      <button type="submit">Submit</button>
-    </form>
-  </body>
-</html>
